Replace jruby-19mode with 1.9.3
Rspec-puppet and jruby seem to be incompatible. When compiling catalogs, rspec-puppet attempts to enforce them to some extent. This is a problem with jruby and execs because it is unable to fork. Not only is it breaking the build, but it's not representative of reality: the jruby environment on the puppetserver only compiles the catalog, the "real" ruby environment on the agent enforces the catalog. By restoring the use of rvm 1.9.3, our build pipeline should be restored.