disable RSpec/NamedSubject for now
this is a complicated cop that isn't easy to fix. It got introduced in
the last rubocop-rspec release. We also use it in a lot of modules so
fixing is time consuming and breaks our modulesync. We already disabled
the cop in https://github.com/voxpupuli/puppet-nodejs/pull/240 which
created the expected result.
lot of modules