Use `YAML.safe_load`, re-enable Rubocop check
This commit replaces calls to `YAML.load` in `spec/spec_helper.rb` with
calls to `YAML.safe_load` and re-enables the related Rubocop check in
the modulesync_config base directory.  Prior to this, because the cop
was not disabled in the module root, modules synced to this repo would
fail Rubocop checks while this modulesync_config repo would pass.  A
change in code was prefered here, because the check is a security check
rather than style, lint, or metrics check, and the code change has no
effect on the outcome of the spec helper.