Portability fixes
Always use either `/bin/sh` or `/usr/bin/env command` as shebang to cope
with distribution-specific path to commands and unavailability of `bash`
on some platforms.

Use `printf(1)` which is portable instead of `echo -n` which is
shell-dependent.
Use a subshell and change directory there instead of relying on pushd /
popd bashisms.