(GH-151) Check for trailing whitespace in Markdown files
This PR can easily be changed to add other files that should be checked
for trailing whitespace.

Disable rubocop cop's in Rakefile that are not actually improving code

Enabling these cops, with regards to the trailing_whitespace task, make
the code more difficult to read and understand. The ability to
understand the code easily should take precedence over any style.