Add a job to warn about modulesync changes
01cfc20164bb4db5432b82d7a59e2227a718b240 added headers, but this isn't
really a good experience for users. It's actually much better to fail in
CI if a diff is detected. That gives them also a way to verify their
changes to .sync.yml are correct.

This takes the approach of checking out the modulesync config with the
version from .msync.yml and running a one off change in offline mode.
Then it uses git diff to see if there are differences. That should fail
the build and also show the actual differences, which makes debugging
easier.