Add script to adjust modules afer dep got migrated
the camptocamp/systemd module is now voxpupuli/systemd. I wrote a script
to update the namespace in .fixtures.yml, the namespace and upper
version in metadata.json and to create a pull request afterwards.

Call it like this:
```
export GITHUB_TOKEN=...
./bin/clean-git-checkouts
./bin/systemd_module_migration.sh
```

This script can be adjusted for other modules. None of our modules
install the systemd module in spec/spec_helper_acceptance.rb nor
spec/acceptance/*, so I didn't implement this.