Limit to known-good setups from acceptance tests
For acceptance testing, we look into metadata.json for supported versions
of Puppet and operating systems, and for each combination try to execute
the acceptance test suites.

However, some combination depend on nonexistent packages, typically for a
recent version of Puppet on an old Linux distribution, or with an old
version of Puppet on a recent Linux distribution.

Instead of letting GitHub compute all possible matrix combinations, rely
on puppet_metadata to build the list of supported combinations.