Add support for local rubocop overrides
All module rubocop configuration is provided by the voxpupuli-test which
provide sensible defaults.  Some modules may however benefit from
overriding some configuration, for example in the LDAP language, a
*distinguished name* is often referred to as *dn* but this 2 letters
variable name will make rubocop choke on the 3 letters minimum for a
variable name test.  A similar issue exist with types and providers
where the Puppet documentation use the *is* and *should* variables to
compare current and expected states, the former variable name not
matching this rule.

Allow fine tunning the rules when to add this kind of exceptions.