Explicitly require YAML in Rakefile
Previously it was relying on other gems to pull in YAML while using it a
bit lower. This makes it explicit which at worst is redundant and at
best prevents failures. It is loaded after github_changelog_generator so
that if the gem isn't present, YAML isn't pulled in either.

Fixes: 9fb2f67c3d80c363106f9d6db718ff8b00f05814