Update REFERENCE.md during CHANGELOG.md creation
At the moment our release rake tasks checks if the REFERENCE.md is
up2date. It doesn't update it. I think it makes sense to generate the
REFERENCE.md in the same step as the CHANGELOG.md, because that will be
reviewed in a PR. This commit implements it. I created
https://github.com/voxpupuli/puppet-logrotate/pull/201 to check if
rubocop is happy.