Rework release preparation tasks
This is a follow-up to https://github.com/voxpupuli/modulesync_config/pull/788

Voxpupuli contributors used to update a module version and `rake
changelog` before openning a PR for preparing a new release.  Some
modules now have a REFERENCE.md file that also needs to be updated (by
running `rake reference`), and forgetting to do so will remained
unnoticed until `rake release` complain that the file is not up-to-date
(so after the release PR was reviewed and merged).

This PR remove the `changelog` and `reference` tasks which are now
part of the voxpupuli-release gem.