anchor most .pmtignore patterns to the root of the module
The PathSpec used by .pmtignore will match all recursive directories
unless there is a "separator" at the start of the pattern per
https://git-scm.com/docs/gitignore.

As an example of the problem, the pattern `log/` will match a dir named
`foo/log`. Prefixing the pattern with `/` forces the pattern to match
only relative to the root of the module. Where as the pattern `/log/`
would not match a dir named `foo/log`.