Allow opting-out of GitHub Action acceptance tests
Some modules acceptance tests cannot be run from a docker container as
exposed by GitHub Actions (e.g. the puppet-virtualbox module want to
install the running kernel headers to build kernel modules (which is not
possible if the kernel of the host system is not one of the kernels
available in the test system), and then load modules in the kernel which
is not allowed.

Add a configuration flag `acceptance_tests` (defaulting to true) and
make modulesync check it before emitting the suitable GitHub Actions for
the module.