CI: run msync with branch foobranch
By default we write into the branch modulesync. That's configured in
modulesync.yml. msync will check for an existing branch and then update
it, instead of master. That's fine for normal operations, but not for
our CI. Our noop run in CI should detect which modulesync_config changes
aren't yet merged, so we need to use a different branch here.