Automatically label modulesync PRs accordingly
When using modulesync, we regularly end-up creating many pull-requests
to merge a `modulesync` branch into main / master.

We have a `modulesync` label that allows such PRs to be excluded from
the ChangeLog… but forgetting to label them when we create them in batch
mean we have to label them one by one by hand which is a PITA.

Adjust the labeller configuration so that PRs against a branch named
`modulesync` is automatically tagged `modulesync`.