Add a utility script to bump supported OS versions
Add support for a newer version of an Operating System if the previous
version is supported.  Update metadata.json, create a feature branch,
commit, and open a PR automatically.

If the previous version was not supported but older version are, prompt
the user interactively for what to do.

If the OS is not supported at all, do nothing.

Typical usage (when Debian 13 is released):

```
bundle exec msync execute -B -- git pull
bundle exec msync execute -B -- $PWD/bin/add-support-for-new-os -o Debian -v 13
```